Remove deny(warnings) from crate config (#28)
While it's useful to disallow warnings in CI, adding `deny(warnings)` to the
crate is actively harmful, as it breaks the code whenever the compiler adds new
warnings.  This has a negative effect on downstream users.  For instance, it's
not currently possible to view the documentation for the current version of the
code, since `cargo doc` fails on a warning about Markdown formatting.

More elaboration can be found here:

https://www.reddit.com/r/rust/comments/f5xpib/psa_denywarnings_is_actively_harmful/